make: control symbol visibility
We seem to be outgrowing export-symbols-regex. On Linux, ld seems to be
quite content linking against hidden symbols, but at least on OSX the
exported symbol visibility is enforced. For instance, exporting only the
"rados_" prefix breaks anything in C++. Unfortunately, export-symbols-regex
doesn't work with C++ due to name mangling.

Enter -fvisibility=hidden, which seems to be what many large projects
are using. Effectively everything is hidden by default, and we
explicitly export what we want.

These are the basics, and there are variants that work on Windows, as
well with more macro magic.

 #define CEPH_EXPORT __attribute__((__visibility__("default")))

 class CEPH_EXPORT ObjectOperation
 {
 public:
   ObjectOperation();
   virtual ~ObjectOperation();

More info:

https://www.gnu.org/software/gnulib/manual/html_node/Exported-Symbols-of-Shared-Libraries.html
http://gcc.gnu.org/wiki/Visibility

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>