ceph, rados.py: change strategies on timeout/interrupt
We run all librados calls in a separate thread so that the main Python
thread can still see its SIGINT indication from ^C and act on it.
Previously we'd set those threads to 'daemon' mode so that Python didn't
wait for them.  Recently, Python has changed its shutdown handling of
daemon threads in a way that's caused this to break.  So instead we
now use a new facility in MonClient to have shutdown() terminate a
stuck authenticate() thread; this way the librados threads don't
need to be marked as daemon threads, but exit normally.

Also, distinguish between interrupt and timeout in run_in_thread(), and
raise specific exceptions for each, so that callers can potentially
change their strategy for handling the issue based on the termination
reason.

Signed-off-by: Dan Mick <dan.mick@redhat.com>