mds: add `dump cache` asok
Alternative to the legacy mds_tell dumpcache.  If no
path is given to asok, send dump in JSON as respond
to command.  This will be a slow for large caches,
so the JSON mode is really only for small test/debug
situations where we want to peer at the cache from
a script.

Signed-off-by: John Spray <john.spray@redhat.com>